static/tests/secure/create-cookie-js.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<script>
  function create_test(origin_page, params, shouldExistInDOM, title) {
    promise_test(t => {
      var name = "create_secure_js_testcookie";
      
      //TODO - make sure all deletes happen from a secure origin
      erase_cookie_from_js(name); //CAUTION
      var value = "" + Math.random();
      document.cookie = name + "=" + value + ";" + params;
    
      //TODO - redirect to origin_page?
      assert_dom_cookie(name, value, shouldExistInDOM);
      
      //cleanup when done
      erase_cookie_from_js(name);

    }, title);
  }

  //
  //Assuming initial origin page is insecure
  //
  
  //basic test
  create_test(INSECURE_ORIGIN, "path=/", true, "insecure origin, root path, no special behavior");
  //You cannot set a Secure cookie via JS from an insecure origin
  create_test(INSECURE_ORIGIN, "path=/;Secure", false, "insecure origin, root path, secure cookie");
  //You cannot set a HTTPOnly cookie via JS (test not needed)
  create_test(INSECURE_ORIGIN, "path=/;HttpOnly", false, "insecure origin, root path, httponly cookie");
  
  //
  //Assuming initial origin page is secure
  //
  
  //basic test
  create_test(SECURE_ORIGIN, "path=/", true, "secure origin, root path, no special behavior");
  //You can set a Secure cookie via JS from a secure origin
  create_test(SECURE_ORIGIN, "path=/;Secure", true, "secure origin, root path, secure cookie");
  //You can set a HTTPOnly cookie via JS from a secure origin, but its only visible via https (test not needed)
  create_test(SECURE_ORIGIN, "path=/;HttpOnly", false, "secure origin, root path, httponly cookie");

</script>
